+++ MORYX-Framework.sln
@@ -216,8 +216,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Drivers", "Drivers", "{93C9
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moryx.Drivers.Mqtt", "src\Moryx.Drivers.Mqtt\Moryx.Drivers.Mqtt.csproj", "{7EDCD56F-EF56-40DF-A25F-5896990D2407}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moryx.Resources.Mqtt.Tests", "src\Tests\Moryx.Recources.Mqtt.Tests\Moryx.Resources.Mqtt.Tests.csproj", "{4C32122A-90B9-4E49-AE3E-985FA5D4AC54}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moryx.Drivers.OpcUa", "src\Moryx.Drivers.OpcUa\Moryx.Drivers.OpcUa.csproj", "{BD3B74DF-9EC2-4A5A-BCFE-1707C83756CC}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moryx.Drivers.OpcUa.Tests", "src\Tests\Moryx.Drivers.OpcUa.Tests\Moryx.Drivers.OpcUa.Tests.csproj", "{CB5BC342-4952-4419-ABB9-0D240684A3DE}"
@@ -307,6 +305,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moryx.Shifts.Management.IntegrationTests", "src\Tests\Moryx.Shifts.Management.IntegrationTests\Moryx.Shifts.Management.IntegrationTests.csproj", "{5E02B439-B91F-4297-9AE4-7B92F2AA2AFF}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moryx.Model.SqlServer", "src\Moryx.Model.SqlServer\Moryx.Model.SqlServer.csproj", "{4402EF2E-CBA8-4EEF-B8A6-EC8364960306}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moryx.Drivers.Mqtt.Tests", "src\Tests\Moryx.Drivers.Mqtt.Tests\Moryx.Drivers.Mqtt.Tests.csproj", "{B2F922C0-F7C3-4C70-9BC5-5B5743BD8D7F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -646,10 +646,6 @@ Global
 		{7EDCD56F-EF56-40DF-A25F-5896990D2407}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7EDCD56F-EF56-40DF-A25F-5896990D2407}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7EDCD56F-EF56-40DF-A25F-5896990D2407}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4C32122A-90B9-4E49-AE3E-985FA5D4AC54}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4C32122A-90B9-4E49-AE3E-985FA5D4AC54}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4C32122A-90B9-4E49-AE3E-985FA5D4AC54}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4C32122A-90B9-4E49-AE3E-985FA5D4AC54}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BD3B74DF-9EC2-4A5A-BCFE-1707C83756CC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BD3B74DF-9EC2-4A5A-BCFE-1707C83756CC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BD3B74DF-9EC2-4A5A-BCFE-1707C83756CC}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -810,6 +806,10 @@ Global
 		{4402EF2E-CBA8-4EEF-B8A6-EC8364960306}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4402EF2E-CBA8-4EEF-B8A6-EC8364960306}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4402EF2E-CBA8-4EEF-B8A6-EC8364960306}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B2F922C0-F7C3-4C70-9BC5-5B5743BD8D7F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B2F922C0-F7C3-4C70-9BC5-5B5743BD8D7F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B2F922C0-F7C3-4C70-9BC5-5B5743BD8D7F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B2F922C0-F7C3-4C70-9BC5-5B5743BD8D7F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -893,7 +893,6 @@ Global
 		{39EDE0BB-F7D8-4E2E-9D1B-36A7A245FA05} = {8BBF84BB-4A8B-49F4-BA74-F25CBEED3F57}
 		{D473D554-26F3-47EC-AE7D-A9754556C9D9} = {8BBF84BB-4A8B-49F4-BA74-F25CBEED3F57}
 		{7EDCD56F-EF56-40DF-A25F-5896990D2407} = {93C94C82-873F-4D31-8A9C-0227047BD5A8}
-		{4C32122A-90B9-4E49-AE3E-985FA5D4AC54} = {0A466330-6ED6-4861-9C94-31B1949CDDB9}
 		{BD3B74DF-9EC2-4A5A-BCFE-1707C83756CC} = {93C94C82-873F-4D31-8A9C-0227047BD5A8}
 		{CB5BC342-4952-4419-ABB9-0D240684A3DE} = {0A466330-6ED6-4861-9C94-31B1949CDDB9}
 		{58CD8D72-1000-479A-B82A-D8C018345239} = {73AFB51B-B1A2-4144-AB4C-9CDE80CBCFBC}
@@ -934,6 +933,7 @@ Global
 		{22358E04-0AE4-4ADA-9018-56C881E313F1} = {8DF13E64-63FC-44A9-A54C-ADEFC356CDE8}
 		{5E02B439-B91F-4297-9AE4-7B92F2AA2AFF} = {8517D209-5BC1-47BD-A7C7-9CF9ADD9F5B6}
 		{4402EF2E-CBA8-4EEF-B8A6-EC8364960306} = {74112169-6672-4907-A187-F055111940A9}
+		{B2F922C0-F7C3-4C70-9BC5-5B5743BD8D7F} = {0A466330-6ED6-4861-9C94-31B1949CDDB9}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {36EFC961-F4E7-49DC-A36A-99594FFB8243}
