+++ src/Tests/Moryx.Drivers.Mqtt.Tests/TestDriverMqttChannelFunctions.cs
@@ -9,11 +9,10 @@ using Microsoft.Extensions.Logging.Abstractions;
 using Moq;
 using Moryx.AbstractionLayer.Products;
 using Moryx.AbstractionLayer.TestTools;
-using Moryx.Drivers.Mqtt;
 using Moryx.Drivers.Mqtt.MqttTopics;
+using Moryx.Drivers.Mqtt.Tests.TestMessages;
 using Moryx.Logging;
 using Moryx.Modules;
-using Moryx.Resources.Mqtt.Tests.TestMessages;
 using Moryx.TestTools.UnitTest;
 using Moryx.Tools;
 using MQTTnet;
@@ -23,7 +22,7 @@ using MQTTnet.Protocol;
 using Newtonsoft.Json;
 using NUnit.Framework;
 
-namespace Moryx.Resources.Mqtt.Tests
+namespace Moryx.Drivers.Mqtt.Tests
 {
     [TestFixture(MqttProtocolVersion.V310)]
     [TestFixture(MqttProtocolVersion.V311)]
