+++ src/Tests/Moryx.Drivers.Mqtt.Tests/TestDriverMqttPrimitiveDatatypeTopic.cs
@@ -8,9 +8,7 @@ using System.Text;
 using System.Threading;
 using Microsoft.Extensions.Logging.Abstractions;
 using Moq;
-using Moryx.AbstractionLayer.Resources;
 using Moryx.AbstractionLayer.TestTools;
-using Moryx.Drivers.Mqtt;
 using Moryx.Drivers.Mqtt.MqttTopics;
 using Moryx.Logging;
 using Moryx.Modules;
@@ -21,7 +19,7 @@ using MQTTnet.Packets;
 using MQTTnet.Protocol;
 using NUnit.Framework;
 
-namespace Moryx.Resources.Mqtt.Tests
+namespace Moryx.Drivers.Mqtt.Tests
 {
     [TestFixture(MqttProtocolVersion.V310)]
     [TestFixture(MqttProtocolVersion.V311)]
