+++ src/Tests/Moryx.Drivers.Mqtt.Tests/TestDriverStateMachine.cs
@@ -7,7 +7,6 @@ using Microsoft.Extensions.Logging.Abstractions;
 using Moq;
 using Moryx.AbstractionLayer.Drivers;
 using Moryx.AbstractionLayer.TestTools;
-using Moryx.Drivers.Mqtt;
 using Moryx.Logging;
 using Moryx.Modules;
 using Moryx.Tools;
@@ -15,7 +14,7 @@ using MQTTnet;
 using MQTTnet.Formatter;
 using NUnit.Framework;
 
-namespace Moryx.Resources.Mqtt.Tests
+namespace Moryx.Drivers.Mqtt.Tests
 {
     [TestFixture(MqttProtocolVersion.V310)]
     [TestFixture(MqttProtocolVersion.V311)]
